+++ core/src/main/java/arez/SpyImpl.java
@@ -241,15 +241,6 @@ final class SpyImpl
    * {@inheritDoc}
    */
   @Override
-  public boolean isScheduled( @Nonnull final Observer observer )
-  {
-    return observer.isScheduled();
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
   public boolean isComputedValue( @Nonnull final ObservableValue<?> observableValue )
   {
     return observableValue.hasOwner();